EDM-310: Move the CA to an async task
1. Move the CA to a singleton
2. Provide initial support for multiple signers/CA identities
3. Move signing to an async task
4. Invoke the async task instead of signing

TODO - SQL queries in the async task are not possible due to
missing functionality in the API. As a result the check for
certificate status is presently done by enumerating all certs.
This is not fir for production and needs to be fixed once
ListParams is extended to allow a NOT NULL query.

Signed-off-by: Anton Ivanov <anton.ivanov@cambridgegreys.com>